requirements.txt
certifi==2018.10.15
dj-database-url==0.5.0
Django==2.1.3
django-crispy-forms==1.7.2
django-isbn-field==0.5.2
envparse==0.2.0
psycopg2-binary==2.7.6.1
python-stdnum==1.10
pytz==2018.7
raven==6.9.0
six==1.11.0
virtualenv==16.1.0
virtualenv-clone==0.4.0
